Validate 'old_text' is not empty before search in EditTool; update error handling in SearchTool; improve error messaging for missing 'items' in TodoTool

+++ src/tools/mod.rs
@@ -825,7 +825,7 @@ mod win32 {
             hNamedPipe: isize,
             lpMode: *mut u32,
             lpMaxCollectionCount: *mut u32,
-            lpCollectDataTimeout: *mut u64,
+            lpCollectDataTimeout: *mut u32,
         ) -> i32;
     }
